requirements_dev.txt
pandas

psycopg2-binary
sql_queries

ipykernel
ipython-sql